Revert all changes related to storing and restoring non-rel data in page server
This includes the following commits:

35a1c3d521729709a7bafe47fb4a3a7bdf52ad65 Specify right LSN in test_createdb.py
d95e1da7425674bec41348066a5950cfc761f043 Fix issue with propagation of CREATE DATABASE to the branch
8465738aa5cd360ae7b71c95e435bd70527f572c [refer #167] Fix handling of pg_filenode.map files in page server
86056abd0e071b86c447a6523226b37160dd5387 Fix merge conflict: set initial WAL position to second segment because of pg_resetwal
2bf2dd1d8871261de4c6a104de5664d32f208260 Add nonrelfile_utils.rs file
20b6279beba10b1f92606e0b4450758eae217c79 Fix restoring non-relational data during compute node startup
06f96f96001ddb0033b23fc6cb7beaa84ce9fa8f Do not transfer WAL to computation nodes: use pg_resetwal for node startup

As well as some older changes related to storing CLOG and MultiXact data as
"pseudorelation" in the page server.

With this revert, we go back to the situtation that when you create a
new compute node, we ship *all* the WAL from the beginning of time to
the compute node. Obviously we need a better solution, like the code
that this reverts. But per discussion with Konstantin and Stas, this
stuff was still half-baked, and it's better for it to live in a branch
for now, until it's more complete and has gone through some review.